Break up utils file
