+++ internal/utils/errors.go
@@ -1,0 +1,9 @@
+package utils
+
+import (
+	"errors"
+)
+
+var (
+	ErrUnknownProvider = errors.New("unknown provider")
+)
