+++ internal/utils/logo.go
@@ -1,0 +1,18 @@
+package utils
+
+var (
+	// Logo for GoDNS.
+	Logo = `
+
+ ██████╗  ██████╗ ██████╗ ███╗   ██╗███████╗
+██╔════╝ ██╔═══██╗██╔══██╗████╗  ██║██╔════╝
+██║  ███╗██║   ██║██║  ██║██╔██╗ ██║███████╗
+██║   ██║██║   ██║██║  ██║██║╚██╗██║╚════██║
+╚██████╔╝╚██████╔╝██████╔╝██║ ╚████║███████║
+ ╚═════╝  ╚═════╝ ╚═════╝ ╚═╝  ╚═══╝╚══════╝
+
+GoDNS V%s
+https://github.com/TimothyYe/godns
+
+`
+)
