+++ internal/utils/settings_test.go
@@ -6,17 +6,6 @@ import (
 	"github.com/TimothyYe/godns/internal/settings"
 )
 
-func TestGetCurrentIP(t *testing.T) {
-	conf := &settings.Settings{IPUrl: "https://myip.biturl.top"}
-	ip, _ := GetCurrentIP(conf)
-
-	if ip == "" {
-		t.Log("IP is empty...")
-	} else {
-		t.Log("IP is:" + ip)
-	}
-}
-
 func TestCheckSettings(t *testing.T) {
 	settingError := &settings.Settings{}
 	if err := CheckSettings(settingError); err == nil {
